added tests for lsi lda, test base
created unified TestBaseModel and aded test for printTopic
